movies_admin/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-debug-toolbar==4.4.6
django-split-settings==1.3.2
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
